+++ src/main/java/org/ptithcm2021/hr_management/model/LeaveBalance.java
@@ -25,10 +25,10 @@ public class LeaveBalance {
     @Column
     private int month;
 
-    private int totalLeaveDay;
-    private int carriedOverDay;
-    private int usedLeaveDay;
-    private int remainingLeaveDay;
+    private double totalLeaveDay;
+    private double carriedOverDay;
+    private double usedLeaveDay;
+    private double remainingLeaveDay;
 
     @ManyToOne
     @JoinColumn(name ="userId")
